pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "aoe2-image-description-generator-for-streamers"
version = "0.1.0"
description = "Generate RU civilization info images for Age of Empires II from aoe2techtree data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "pillow>=10.0",
  "pyyaml>=6.0",
]

[project.scripts]
aoe2civgen = "aoe2civgen.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aoe2civgen"]

[tool.uv]
package = true
